+++ weed/util/chunk_cache/chunk_cache.go
@@ -1,4 +1,4 @@
-package pb_cache
+package chunk_cache
 
 import (
 	"time"
